requirements.txt
drf-yasg==1.21.3
dj-database-url==1.0.0
Django==4.0.4
django-dotenv==1.4.2
django-extensions==3.1.5
django-heroku==0.3.1
django-rest-authtoken==2.1.4
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
django-filter==22.1